remove stale html file
so that the changes made to the rmd file aren't ignored by the server. Otherwise, it appears that the server will read the html file instead of the changed RMD file. And the build website action doesn't seem to create a new html file if there already exists an html file.